postgres_to_es/pyproject.toml
[tool.poetry]
name = "postgres-to-es"
version = "0.0.1"
description = "Postgres to Elasticsearch ETL"
authors = ["shitcoding <shitcoder@pm.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
loguru = "^0.7.2"
psycopg2-binary = "2.9.9"
backoff = "2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
